Dialogue System In Place
The dialogue system is in place and it works.

However, currently it's limited to three options due to the constraint on UI. If necessary it's an easy fix to expand on size, we'll just need to discuss on the layout of UI.

As of now, I have only thought of changing the background image, but if you want to have sound effects, expression change, and animation we'll have to talk about when and how things occur before I can implement those systems.